pyproject.toml
[project]
name = "ghstats2"
version = "0.1.0"
description = "GitHub traffic statistics collector for NatLabRockies repositories"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.5",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "polars>=0.20",
    "pyarrow>=15.0",
    "click>=8.1",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "respx>=0.21",
    "ruff>=0.3",
]
notebook = [
    "jupyter>=1.0",
    "plotly>=5.18",
]

[project.scripts]
ghstats2 = "ghstats2.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ghstats2"]

[tool.ruff]
src = ["src"]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "D",   # pydocstring
]
ignore = ["D100", "D104", "D105"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
